requirements.txt
Package       Version  
------------- ---------
certifi       2020.12.5
chardet       4.0.0    
click         8.0.1    
Flask         2.0.1    
idna          2.10     
itsdangerous  2.0.1    
Jinja2        3.0.1    
MarkupSafe    2.0.1    
passlib       1.7.4    
pip           20.0.2   
pkg-resources 0.0.0    
pymongo       3.11.4   
python-dotenv 0.17.1   
requests      2.25.1   
setuptools    44.0.0   
urllib3       1.26.5   
Werkzeug      2.0.1    
